services/booking-service/requirements.txt
fastapi
uvicorn
google-api-python-client
psycopg2-binary
python-dotenv
